+++ app/Http/Controllers/TagsController.php
@@ -96,6 +96,14 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        //? Before deleting tag, check if the tag dont got a posts attach to it!
+        if ($tag->posts->count() > 0){
+
+            session()->flash('error', 'Tag cannot been deleted, because it is associated to some posts!');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
